+++ src/main/java/dev/krk/emulator/chipsekiz/opcodes/Op3XNN.java
@@ -5,7 +5,7 @@ package dev.krk.emulator.chipsekiz.opcodes;
  * Skips the next instruction if VX equals NN.
  */
 public class Op3XNN extends OpPXNN {
-    public Op3XNN(int vx, int imm) {
+    public Op3XNN(int vx, byte imm) {
         super(3, vx, imm);
     }
 }
