+++ src/main/java/dev/krk/emulator/chipsekiz/opcodes/Op6XNN.java
@@ -5,7 +5,7 @@ package dev.krk.emulator.chipsekiz.opcodes;
  * Sets VX to NN.
  */
 public class Op6XNN extends OpPXNN {
-    public Op6XNN(int vx, int imm) {
+    public Op6XNN(int vx, byte imm) {
         super(6, vx, imm);
     }
 }
