+++ src/main/java/dev/krk/emulator/chipsekiz/opcodes/Op7XNN.java
@@ -2,7 +2,7 @@ package dev.krk.emulator.chipsekiz.opcodes;
 
 
 public class Op7XNN extends OpPXNN {
-    public Op7XNN(int vx, int imm) {
+    public Op7XNN(int vx, byte imm) {
         super(7, vx, imm);
     }
 }
