+++ src/main/java/dev/krk/emulator/chipsekiz/opcodes/OpCXNN.java
@@ -2,7 +2,7 @@ package dev.krk.emulator.chipsekiz.opcodes;
 
 
 public class OpCXNN extends OpPXNN {
-    public OpCXNN(int vx, int imm) {
+    public OpCXNN(int vx, byte imm) {
         super(0xC, vx, imm);
     }
 }
